% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.class.R
\name{check.class}
\alias{check.class}
\title{Function to check whether obj belongs to class or
is a list of class (if list.of.class==TRUE)}
\usage{
check.class(obj, class, list.of.class = TRUE)
}
\description{
Function to check whether obj belongs to class or
is a list of class (if list.of.class==TRUE)
}
\keyword{internal}
