% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.ss.bim.R
\name{comp.ss.bim}
\alias{comp.ss.bim}
\title{Compare summary statistics and PLINK .bim data.frames}
\usage{
comp.ss.bim(ss.df, bim.df, exclude.ambiguous = F)
}
\arguments{
\item{ss.df}{A data.frame of summary statistics}

\item{bim.df}{a data.frame from A PLINK .bim file}

\item{exclude.ambiguous}{should ambiguous SNPs (i.e. A/T or G/C) be excluded}
}
\value{
A list with 3 vectors 
\item{ss.order}{the order the SNPs in the summary statistics must be rearranged in}
\item{bim.extract}{the SNPs to be extracted from the PLINK bfile, i.e. that 
are in the summary statistics data.frame}
\item{rev}{A vector of 1 and -1 to facilitate the flipping SNPs effects for SNPs whose
reference and alternative alleles are reversely coded between the two data.frames}
}
\description{
Compare summary statistics and PLINK .bim data.frames
}
\details{
Both \code{ss.df} and \code{bim.df} must be \code{data.frame}s with exactly 
three columns in this order: (1) SNP ids, (2) reference allele, (3) alternative allele
}
\keyword{#@export}
\keyword{internal}
