% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countlines}
\alias{countlines}
\title{Count number of lines in a text file}
\usage{
countlines(fileName)
}
\arguments{
\item{fileName}{Name of file}
}
\description{
Count number of lines in a text file
}
\keyword{internal}
