% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.beta.R
\name{cp.beta}
\alias{cp.beta}
\title{Get the "overall" beta after cross-prediction}
\usage{
cp.beta(cp.plink.linear, cp.lassosum, method = c("Method1", "Method2"),
  load = "", save = "", force = FALSE)
}
\arguments{
\item{cp.plink.linear}{An object returned from cp.plink.linear()}

\item{cp.lassosum}{An object returned from cp.lassosum()}

\item{method}{Either the Method1 or Method2 Polygenic Score}

\item{load}{Load the pseudoinverse}

\item{save}{Save the pseudoinverse (in armadillo binary format)}
}
\value{
A vector of betas
}
