% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.lassosum.list.R
\name{cp.lassosum.list}
\alias{cp.lassosum.list}
\title{For a list of cp.plink.linear objects}
\usage{
cp.lassosum.list(cp.plink.linear, trace = 0, keep.ref = NULL, ...)
}
\description{
For a list of cp.plink.linear objects
}
\keyword{internal}
