% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.lassosum.validate.R
\name{cp.lassosum.validate}
\alias{cp.lassosum.validate}
\title{For a list of cp.plink.linear objects}
\usage{
cp.lassosum.validate(list.of.lpipe, cp.plink.linear, Method2, pseudovalidation,
  scale = TRUE, plot, validate.function, cluster, trace, details,
  list.of.bfiles)
}
\description{
For a list of cp.plink.linear objects
}
\keyword{internal}
