% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.plink.linear.R
\name{cp.plink.linear}
\alias{cp.plink.linear}
\title{Generate summary statistics for cross-prediction}
\usage{
cp.plink.linear(bfile, nfolds = 5, fold = NULL, pheno = NULL,
  covar = NULL, keep = NULL, remove = NULL, extract = NULL,
  exclude = NULL, chr = NULL, force = FALSE, fast = TRUE, ...)
}
\arguments{
\item{nfolds}{Number of folds}

\item{fold}{A vector of fold ID}

\item{pheno}{A vector of phenotype}

\item{covar}{A vector/matrix of covariates}

\item{keep, remove, extract, exclude, chr}{see \code{\link{parseselect}}}

\item{force}{Force}

\item{fast}{see details}
}
\value{
A \code{cp.plink.linear} object with the following elements
\item{cor}{A list of correlation coefficients by folds}
\item{chr, pos, A1, snp}{Variant identifier}
\item{pheno.by.fold}{A list of phenotypes by fold}
\item{covar.by.fold}{A list of covariates by fold}
\item{keep}{A logical vector of which sample were included. Alternatively, \code{NULL} means all sample}
\item{extract}{A logical vector of which variant/SNP were used. Alternatively, \code{NULL} means all variants.} 
\item{n}{Number of samples used}
\item{p}{Number of variants included}
\item{nonmiss}{A list of vectors of number of non-missing samples used by fold}
\item{bfile}{The bfile used}
}
\details{
If fast == TRUE, the summary statistics are calculated for 
         each fold, and then combined together by averaging. For example,
         assume there are 5 folds, then for fold 1, the summary stats
         for fold 2, 3, 4, and 5 are averaged. If fast == FALSE, 
         we actually run plink using the entire dataset minus the fold 1, 
         and repeat this process for folds 2, 3, 4, and 5.
}
