% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.sumstats.R
\name{cp.sumstats}
\alias{cp.sumstats}
\title{Function to convert external fold-specific summary statistics}
\usage{
cp.sumstats(cor, chr = NULL, pos = NULL, snp = NULL, A1 = NULL,
  A2 = NULL, fold = NULL, nonmiss = NULL, bfile = NULL,
  exclude.ambiguous = TRUE, fast = TRUE)
}
\keyword{internal}
