% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elnet}
\alias{elnet}
\title{Performs elnet}
\usage{
elnet(lambda1, lambda2, diag, X, r, thr, x, yhat, trace, maxiter)
}
\arguments{
\item{lambda1}{lambda}

\item{lambda2}{lambda}

\item{X}{genotype Matrix}

\item{r}{correlations}

\item{thr}{threshold}

\item{x}{beta coef}

\item{yhat}{A vector}

\item{trace}{if >1 displays the current iteration}

\item{maxiter}{maximal number of iterations}
}
\value{
conv
}
\description{
Performs elnet
}
\keyword{internal}
