% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.blocks.R
\name{group.blocks}
\alias{group.blocks}
\title{Group blocks into chunks so as not to exhaust memory or hit 
matrix size ceiling}
\usage{
group.blocks(Blocks, parseselect, mem.limit = 2^32 * 8, chunks = NULL,
  cluster = NULL)
}
\arguments{
\item{parseselect}{An object returned by parseselect()}
}
\description{
Group blocks into chunks so as not to exhaust memory or hit 
matrix size ceiling
}
\keyword{internal}
