% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassosum.R, R/package_lassosum.R
\docType{package}
\name{lassosum}
\alias{lassosum}
\alias{lassosum-package}
\title{Function to obtain LASSO estimates of a regression problem given summary statistics
and a reference panel}
\usage{
lassosum(cor, bfile, lambda = exp(seq(log(0.001), log(0.1), length.out = 20)),
  shrink = 0.9, thr = 1e-04, init = NULL, trace = 0, maxiter = 10000,
  keep = NULL, extract = NULL, exclude = NULL, remove = NULL,
  chr = NULL)
}
\arguments{
\item{cor}{A vector of correlations (\eqn{r})}

\item{bfile}{PLINK bfile (as character, without the .bed extension)}

\item{lambda}{A vector of \eqn{\lambda}s (the tuning parameter)}

\item{shrink}{The shrinkage parameter \eqn{s} for the correlation matrix \eqn{R}}

\item{thr}{convergence threshold for \eqn{\beta}}

\item{init}{Initial values for \eqn{\beta} as a vector of the same length as \code{cor}}

\item{trace}{An integer controlling the amount of output generated.}

\item{maxiter}{Maximum number of iterations}

\item{keep}{samples to keep}

\item{extract}{SNPs to extract}

\item{exclude}{SNPs to exclude}

\item{chr}{a vector of chromosomes}

\item{samples}{to remove}
}
\value{
A list with the following
\item{lambda}{same as the lambda input}
\item{beta}{A matrix of estimated coefficients}
\item{conv}{A vector of convergence indicators. 1 means converged. 0 not converged.}
\item{pred}{\eqn{=\sqrt(1-s)X\beta}}
\item{loss}{\eqn{=(1-s)\beta'X'X\beta/n - 2\beta'r}}
\item{fbeta}{\eqn{=\beta'R\beta - 2\beta'r + 2\lambda||\beta||_1}}
\item{sd}{The standard deviation of the reference panel SNPs}
\item{shrink}{same as input}
\item{nparams}{Number of non-zero coefficients}
}
\description{
The lassosum package provides three categories of important functions:
lassosum, bar and baz.
}
\details{
A function to find the minimum of \eqn{\beta} in  
\deqn{f(\beta)=\beta'R\beta - 2\beta'r + 2\lambda||\beta||_1}
where 
\deqn{R=(1-s)X'X/n + sI}
is a shrunken correlation matrix, with \eqn{X} being standardized reference panel.
\eqn{s} should take values in (0,1]. \eqn{r} is a vector of correlations. 
\code{keep}, \code{remove} could take one of three 
formats: (1) A logical vector indicating which indivduals to keep/remove, 
(2) A \code{data.frame} with two columns giving the FID and IID of the indivdiuals
to keep/remove (matching those in the .fam file), or (3) a character scalar giving the text file with the FID/IID. 
Likewise \code{extract}, \code{exclude} can also take one of the three formats,
except with the role of the FID/IID data.frame replaced with a character vector of 
SNP ids (matching those in the .bim file).
}
\note{
Missing genotypes are interpreted as having the homozygous A2 alleles in the 
PLINK files (same as the \code{--fill-missing-a2} option in PLINK).
}
\section{Foo functions}{

The lassosum functions ...
}

