% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassosumR.R
\name{lassosumR}
\alias{lassosumR}
\title{Function to obtain LASSO estimates of a regression problem given summary statistics
and a reference panel (without PLINK bfile)}
\usage{
lassosumR(cor, refpanel, lambda = exp(seq(log(0.001), log(0.1),
  length.out = 20)), shrink = 0.9, ridge = F, thr = 1e-04,
  init = NULL, trace = 0, maxiter = 10000, blocks = NULL)
}
\arguments{
\item{cor}{A vector of correlations (\eqn{r})}

\item{refpanel}{reference panel as \code{data.frame} or \code{matrix}}

\item{lambda}{A vector of \eqn{\lambda}s (the tuning parameter)}

\item{shrink}{The shrinkage parameter \eqn{s} for the correlation matrix \eqn{R}}

\item{ridge}{Produce ridge regression results also (slow if nrow(refpanel) > 2000)}

\item{thr}{convergence threshold for \eqn{\beta}}

\item{init}{Initial values for \eqn{\beta}}

\item{trace}{An integer controlling the amount of output generated.}

\item{maxiter}{Maximum number of iterations}

\item{blocks}{A vector to split the genome by blocks (coded as c(1,1,..., 2, 2, ..., etc.))}
}
\value{
A list with the following
\item{lambda}{same as the lambda input}
\item{beta}{A matrix of estimated coefficients}
\item{conv}{A vector of convergence indicators. 1 means converged. 0 not converged.}
\item{pred}{\eqn{=(1-s)X\beta}}
\item{loss}{\eqn{=(1-s)\beta'X'X\beta/n - 2\beta'r}}
\item{fbeta}{\eqn{=\beta'R\beta - 2\beta'r + 2\lambda||\beta||_1}}
\item{sd}{The standard deviation of the reference panel SNPs}
\item{shrink}{same as input}
\item{nparams}{Number of non-zero coefficients}
\item{ridge}{ridge regression estimates}
}
\description{
Function to obtain LASSO estimates of a regression problem given summary statistics
and a reference panel (without PLINK bfile)
}
\details{
A function to find the minimum of \eqn{\beta} in  
\deqn{f(\beta)=\beta'R\beta - 2\beta'r + 2\lambda||\beta||_1}
where 
\deqn{R=(1-s)X'X/n + sI}
is a shrunken correlation matrix, with \eqn{X} being standardized reference panel.
\eqn{s} should take values in (0,1]. \eqn{r} is a vector of correlations.
}
\note{
\itemize{
\item Missing values in \code{refpanel} are filled with 0. 
\item Unlike lassosum, we do not provide the options keep/remove/extract/exclude.
It is thus up to the user to ensure the SNPs in the reference panel corresponds 
to those in the correlations.
}
}
\keyword{#@export}
\keyword{internal}
