% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.vector.R
\name{logical.vector}
\alias{logical.vector}
\title{Function to create a logical vector based on position and
size, i.e. total length of vector}
\usage{
logical.vector(positions, size)
}
\arguments{
\item{positions}{Positions to extract}

\item{size}{Length of logical vector}
}
\description{
It's basically the reverse of base::which()
}
