% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchpos.R
\name{matchpos}
\alias{matchpos}
\title{Function to match a set of variants to a reference
by chromosome and/or position and/or SNP/rsIDs}
\usage{
matchpos(tomatch, ref.df, auto.detect.tomatch = T, auto.detect.ref = T,
  chr = "", ref.chr = "", pos = "", ref.pos = "", snp = "",
  ref.snp = "", ref = "", ref.ref = "", alt = "", ref.alt = "",
  exclude.ambiguous = F, silent = F, rm.duplicates = F)
}
\arguments{
\item{tomatch}{A data.frame or data.table}

\item{ref.df}{A data.frame or data.table}

\item{rm.dumplicates}{Remove SNPs with more than one match}
}
\description{
If REF/ALT alleles available, will also check them for reverse coding. 
Specify the exclude.ambiguous option for possible sense/antisense ambiguity.
}
\keyword{internal}
