% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiBed3}
\alias{multiBed3}
\title{imports genotypeMatrix and multiplies it by a constant}
\usage{
multiBed3(fileName, N, P, input, col_skip_pos, col_skip, keepbytes, keepoffset)
}
\arguments{
\item{fileName}{location of bam file}

\item{N}{number of subjects}

\item{P}{number of positions}

\item{input}{the constant}

\item{col_skip_pos}{which variants should we skip}

\item{col_skip}{which variants should we skip}

\item{keepbytes}{which bytes to keep}

\item{keepoffset}{what is the offset}
}
\value{
an armadillo genotype matrix
}
\description{
imports genotypeMatrix and multiplies it by a constant
}
\keyword{internal}

