% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiBed3sp}
\alias{multiBed3sp}
\title{Multiply genotypeMatrix by a matrix (sparse)}
\usage{
multiBed3sp(fileName, N, P, beta, nonzeros, colpos, ncol, col_skip_pos,
  col_skip, keepbytes, keepoffset, trace)
}
\arguments{
\item{fileName}{location of bam file}

\item{N}{number of subjects}

\item{P}{number of positions}

\item{col_skip_pos}{which variants should we skip}

\item{col_skip}{which variants should we skip}

\item{keepbytes}{which bytes to keep}

\item{keepoffset}{what is the offset}

\item{input}{the matrix}
}
\value{
an armadillo genotype matrix
}
\description{
Multiply genotypeMatrix by a matrix (sparse)
}
\keyword{internal}
