% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol.bfile.R
\name{ncol.bfile}
\alias{ncol.bfile}
\title{Obtains the number of column (SNPs) in a PLINK bfile}
\usage{
ncol.bfile(bfile)
}
\arguments{
\item{bfile}{Plink file stem}
}
\value{
an integer with the number of columns
#@keywords internal
}
\description{
Obtains the number of column (SNPs) in a PLINK bfile
}
