% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize}
\alias{normalize}
\title{normalize genotype matrix}
\usage{
normalize(genotypes)
}
\arguments{
\item{genotypes}{a armadillo genotype matrix}
}
\value{
standard deviation
}
\description{
normalize genotype matrix
}
\keyword{internal}
