% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.a.file.R
\name{not.a.file}
\alias{not.a.file}
\title{A function to protect a single SNP id from being interpreted as a file}
\usage{
not.a.file(scalar_character)
}
\description{
A function to protect a single SNP id from being interpreted as a file
}
