% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.bfile.R
\name{nrow.bfile}
\alias{nrow.bfile}
\title{Obtains the number of individuals in a PLINK bfile}
\usage{
nrow.bfile(bfile)
}
\arguments{
\item{bfile}{Plink file stem}
}
\value{
an integer with the number of rows
#@keywords internal
}
\description{
Obtains the number of individuals in a PLINK bfile
}
