% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.by.fold.R
\name{organize.by.fold}
\alias{organize.by.fold}
\title{lassosum.pipeline objects, orginally organized by fold then by chromosome, 
are merged across chromosome}
\usage{
organize.by.fold(list.of.list.of.lp)
}
\arguments{
\item{list.of.list.of.lp}{A list of list of \code{lassosum.pipeline} objects
from the output of \code{cp.lassosum}, when
\code{list.of.lpipe.output=TRUE}}
}
\description{
lassosum.pipeline objects, orginally organized by fold then by chromosome, 
are merged across chromosome
}
