% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overallbeta}
\alias{overallbeta}
\title{An overall beta for cross-prediction}
\usage{
overallbeta(fileName, N, P, col_skip_pos, col_skip, keepbytes, keepoffset, pred,
  meanbeta, save = "", load = "")
}
\arguments{
\item{fileName}{location of bam file}

\item{N}{number of subjects}

\item{P}{number of positions}

\item{col_skip_pos}{which variants should we skip}

\item{col_skip}{which variants should we skip}

\item{keepbytes}{which bytes to keep}

\item{keepoffset}{what is the offset}

\item{pred}{The expected prediction}

\item{meanbeta}{mean beta in the cross-prediction}
}
\value{
an armadillo vector
}
\description{
An overall beta for cross-prediction
}
\keyword{internal}
