% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2cor.R
\name{p2cor}
\alias{p2cor}
\title{Function to convert p-values to correlation via the t-statistic}
\usage{
p2cor(p, n, sign = rep(1, length(p)))
}
\arguments{
\item{p}{Vector of p-values}

\item{n}{Sample size}

\item{sign}{A vector giving the sign of the correlations (e.g. the log odds ratios)}
}
\value{
A vector of correlations
}

