% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseselect.R
\name{parseselect}
\alias{parseselect}
\title{Parse the keep/remove/extract/exclude/chr options}
\usage{
parseselect(bfile, extract = NULL, exclude = NULL, keep = NULL,
  remove = NULL, chr = NULL, export = FALSE)
}
\arguments{
\item{bfile}{plink file stem}

\item{extract}{SNPs to extract}

\item{exclude}{SNPs to exclude}

\item{keep}{samples to keep}

\item{remove}{samples to remove}

\item{chr}{a vector of chromosomes}
}
\value{
a list with
\item{keep}{Either NULL or a logical vector of which individuals to keep}
\item{extract}{Either NULL or a logical vector of which SNPs to extract}
\item{N}{Number of rows in the PLINK bfile}
\item{P}{Number of columns in the PLINK bfile}
\item{n}{Number of rows in the PLINK bfile after keep}
\item{p}{Number of columns in the PLINK bfile after extract}
}
\details{
\code{keep}, \code{remove} could take one of three 
formats: (1) A logical vector indicating which indivduals to keep/remove, 
(2) A \code{data.frame} with two columns giving the FID and IID of the indivdiuals
to keep/remove (matching those in the .fam file), or (3) a character scalar giving 
the text file with the FID/IID. Note that these files should have no headers. 
Likewise \code{extract}, \code{exclude} can also take one of the three formats,
except with the role of the FID/IID data.frame replaced with a character vector of 
SNP ids (matching those in the .bim file).
}
\keyword{internal}
