% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs.R
\name{pgs}
\alias{pgs}
\title{Computes polygenic scores as a genotype matrix multiplied by a matrix of weights}
\usage{
pgs(bfile, weights, keep = NULL, extract = NULL, exclude = NULL,
  remove = NULL, chr = NULL, cluster = NULL)
}
\arguments{
\item{bfile}{A plink bfile stem}

\item{weights}{The weights for the SNPs (\eqn{\beta})}

\item{keep}{samples to keep (see \code{\link{parseselect}})}

\item{extract}{SNPs to extract (see \code{\link{parseselect}})}

\item{exclude}{SNPs to exclude (see \code{\link{parseselect}})}

\item{remove}{samples to remove (see \code{\link{parseselect}})}

\item{chr}{a vector of chromosomes}

\item{cluster}{A \code{cluster} object from the \code{parallel} package. 
For parallel processing.}
}
\value{
A matrix of Polygenic Scores
}
\details{
A function to calculate \eqn{X\beta} where \eqn{X} is the genotype matrix
in the plink bfile.
}
\note{
\itemize{
\item Missing genotypes are interpreted as having the homozygous A2 alleles in the 
PLINK files (same as the \code{--fill-missing-a2} option in PLINK). 
\item The number of rows in \code{weights} should be the same as the number of
SNPs in the bfile after extract/exclude/chr.
}
}

