% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.validate.lassosum.R
\name{plot.validate.lassosum}
\alias{plot.validate.lassosum}
\title{Plot function for \code{validate.lassosum} objects}
\usage{
\method{plot}{validate.lassosum}(obj, legend.x = "bottomright",
  log = "x")
}
\description{
Plot function for \code{validate.lassosum} objects
}
