% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudovalidate.R,
%   R/pseudovalidate.lassosum.pipeline.R
\name{pseudovalidate}
\alias{pseudovalidate}
\alias{pseudovalidate.lassosum.pipeline}
\title{Function to perform pseudovalidation from a lassosum.pipeline object}
\usage{
pseudovalidate(...)

\method{pseudovalidate}{lassosum.pipeline}(ls.pipeline,
  test.bfile = NULL, keep = NULL, remove = NULL, trace = 1,
  destandardize = F, plot = T, exclude.ambiguous = T,
  cluster = NULL, rematch = !is.null(test.bfile), ...)
}
\arguments{
\item{...}{parameters to pass to \code{\link{pseudovalidation}}}

\item{ls.pipeline}{A lassosum.pipeline object}

\item{test.bfile}{The (\href{https://www.cog-genomics.org/plink2/formats#bed}{PLINK bfile} for the test dataset}

\item{keep}{Participants to keep (see \code{\link{lassosum}} for more details)}

\item{remove}{Participants to remove}

\item{trace}{Controls amount of output}

\item{destandardize}{Should coefficients from \code{\link{lassosum}} be 
destandardized using test dataset standard deviations before being returned?}

\item{plot}{Should the validation plot be plotted?}

\item{exclude.ambiguous}{Should ambiguous SNPs (C/G, A/T) be excluded?}

\item{cluster}{A \code{cluster} object from the \code{parallel} package for parallel computing}

\item{rematch}{Forces a rematching of the ls.pipline beta's with the new .bim file}
}
\description{
Function to perform pseudovalidation from a lassosum.pipeline object
}
\details{
Pseudovalidation is explained in Mak et al (2016). It helps 
choosing a value of \code{lambda} and \code{s} in the absence of a validation
phenotype.
}
