% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudovalidation.R
\name{pseudovalidation}
\alias{pseudovalidation}
\title{Performs `pseudovalidation' to select the best \eqn{\lambda} value in lassosum}
\usage{
pseudovalidation(bfile, beta, cor, sd = NULL, keep = NULL,
  extract = NULL, exclude = NULL, remove = NULL, chr = NULL,
  cluster = NULL, ...)
}
\arguments{
\item{bfile}{A plink bfile stem}

\item{beta}{The matrix of estimated \eqn{\beta}s}

\item{cor}{The vector of correlations (\eqn{r})}

\item{sd}{The standard deviation of the SNPs}

\item{keep}{samples to keep}

\item{extract}{SNPs to extract}

\item{exclude}{SNPs to exclude}

\item{remove}{samples to remove}

\item{chr}{a vector of chromosomes}

\item{cluster}{A \code{cluster} object from the \code{parallel} package for parallel computing}
}
\value{
the results of the pseudovalidation, i.e. \eqn{f(\lambda)}
}
\description{
Performs `pseudovalidation' to select the best \eqn{\lambda} value in lassosum
}
\details{
A function to calculate  
\deqn{f(\lambda)=\beta'r/\sqrt{\beta'X'X\beta}} 
where \eqn{X} is the standardized genotype matrix divided by \eqn{\sqrt n}, 
and \eqn{r} is a vector of (shrunken) correlations.
}
\note{
\itemize{
\item Missing genotypes are interpreted as having the homozygous A2 alleles in the 
PLINK files (same as the \code{--fill-missing-a2} option in PLINK). 
\item The number of rows in \code{beta} and the length of \code{cor} should be the same as the number of
SNPs in the bfile after extract/exclude/chr.
}
}
\keyword{internal}
