% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.thresh.old.R
\name{pval.thresh.old}
\alias{pval.thresh.old}
\title{Fast way to do p-value thresholding (without looping over the thresholds)}
\usage{
pval.thresh.old(pvals, p.thresholds, beta, bfile, keep = NULL,
  remove = NULL, extract = NULL, exclude = NULL, chr = NULL,
  mem.limit = 4 * 10^9, chunks = NULL, cluster = NULL, trace = 0)
}
\description{
Fast way to do p-value thresholding (without looping over the thresholds)
}
\keyword{internal}
