% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table2.R
\name{read.table2}
\alias{read.table2}
\title{Function to read a text file}
\usage{
read.table2(file, header = F, data.table = F, check.names = TRUE,
  ...)
}
\description{
Function to read a text file
}
\keyword{internal}
