% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbfile.R
\name{readbfile}
\alias{readbfile}
\title{read a PLINK bfile file into a matrix}
\usage{
readbfile(bfile, keep = NULL, extract = NULL, exclude = NULL,
  remove = NULL, chr = NULL, fillmissing = F)
}
\arguments{
\item{bfile}{PLINK bfile (as character, without the .bed extension)}

\item{keep}{SNPs to keep}

\item{extract}{samples to extract}

\item{exclude}{samples to exclude}

\item{remove}{SNPs to remove}

\item{chr}{a vector of chromosomes}

\item{fillmissing}{Whether to fill missing values with 0}
}
\value{
A genotype matrix of 0,1, and 2 (possibly with NaNs)
}

