% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.study.effects.R
\name{remove.study.effects}
\alias{remove.study.effects}
\title{Function to remove an individual study's effect contribution to a 
fixed-effect meta-analysis summary statistics}
\usage{
remove.study.effects(effectsize, se, meta.effectsize, meta.se)
}
\description{
Function to remove an individual study's effect contribution to a 
fixed-effect meta-analysis summary statistics
}
\keyword{internal}
