% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runElnet}
\alias{runElnet}
\title{Runs elnet with various parameters}
\usage{
runElnet(lambda, shrink, fileName, r, N, P, col_skip_pos, col_skip,
  keepbytes, keepoffset, thr, x, trace, maxiter, startvec, endvec)
}
\arguments{
\item{fileName}{the file name of the reference panel}

\item{r}{a vector of correlations}

\item{N}{number of subjects}

\item{P}{number of position in reference file}

\item{thr}{threshold}

\item{x}{a numeric vector of beta coefficients}

\item{trace}{if >1 displays the current iteration}

\item{maxiter}{maximal number of iterations}

\item{lambda1}{a vector of lambdas (lambda2 is 0)}

\item{col_skip_posR}{which variants should we skip}

\item{col_skipR}{which variants should we skip}

\item{keepbytesR}{required to read the PLINK file}

\item{keepoffsetR}{required to read the PLINK file}

\item{Constant}{a constant to multiply the standardized genotype matrix}
}
\value{
a list of results
}
\description{
Runs elnet with various parameters
}
\keyword{internal}
