% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.bfile.R
\name{sd.bfile}
\alias{sd.bfile}
\title{Obtain the SNP-wise standard deviations from the PLINK bfile}
\usage{
sd.bfile(bfile, keep = NULL, remove = NULL, extract = NULL,
  exclude = NULL, chr = NULL, trace = 0, ...)
}
\arguments{
\item{bfile}{plink file stem}

\item{keep}{samples to keep}

\item{remove}{samples to remove}

\item{extract}{SNPs to extract}

\item{exclude}{SNPs to exclude}

\item{chr}{a vector of chromosomes}
}
\description{
Obtain the SNP-wise standard deviations from the PLINK bfile
}
\details{
\code{keep}, \code{remove} could take one of three 
formats: (1) A logical vector indicating which indivduals to keep/remove, 
(2) A \code{data.frame} with two columns giving the FID and IID of the indivdiuals
to keep/remove (matching those in the .fam file), or (3) a character scalar giving the text file with the FID/IID. 
Likewise \code{extract}, \code{exclude} can also take one of the three formats,
except with the role of the FID/IID data.frame replaced with a character vector of 
SNP ids (matching those in the .bim file).
}
\note{
Missing genotypes are interpreted as having the homozygous A2 alleles in the 
PLINK files (same as the \code{--fill-missing-a2} option in PLINK).
}
\keyword{internal}
