% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectregion.R
\name{selectregion}
\alias{selectregion}
\title{Internal function to parse extract}
\usage{
selectregion(keep)
}
\arguments{
\item{keep}{A boolean vector of which position to keep}
}
\description{
Internal function to parse extract
}
\keyword{internal}
