% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitgenome.R
\name{splitgenome}
\alias{splitgenome}
\title{Function to split a set of SNPs by their position using a reference
Break points should be given for the reference by chromosome}
\usage{
splitgenome(CHR, POS, ref.CHR, ref.breaks, details = T, right = TRUE)
}
\description{
Function to split a set of SNPs by their position using a reference
Break points should be given for the reference by chromosome
}
\keyword{internal}
