% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitvec.from.bfile.R
\name{splitvec.from.bfile}
\alias{splitvec.from.bfile}
\title{Function to create split vectors from the info in vectors of bfiles}
\usage{
splitvec.from.bfile(bfile)
}
\description{
Function to create split vectors from the info in vectors of bfiles
}
\keyword{internal}
