% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.lassosum.pipeline.R
\name{subset.lassosum.pipeline}
\alias{subset.lassosum.pipeline}
\title{Subset a lassosum.pipeline object by lambda and s}
\usage{
\method{subset}{lassosum.pipeline}(lassosum.pipeline, s = NULL,
  lambda = NULL)
}
\arguments{
\item{lassosum.pipeline}{An object returned by lassosum.pipeline}

\item{s}{Value(s) of s to restrict to}

\item{lambda}{Value(s) of lambda to restrict to}
}
\value{
A lassosum.pipeline object
}
\description{
Subset a lassosum.pipeline object by lambda and s
}
\details{
This function is usually used to reapply a validated pgs to a new data.set. 
See example below.
}
\examples{
\dontrun{
 ### Run lassosum using standard pipeline ### 
 lp <- lassosum.pipeline(cor=cor, chr=ss$Chr, pos=ss$Position, 
                          A1=ss$A1, A2=ss$A2,
                          ref.bfile=ref.bfile, test.bfile=test.bfile, 
                          LDblocks = ld)
 v <- validate(lp)
 lp2 <- subset(lp, s=v$best.s, lambda=v$best.lambda)
 v2 <- validate(lp2)
}
}
