% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R, R/validate.lassosum.pipeline.R
\name{validate}
\alias{validate}
\alias{validate.lassosum.pipeline}
\title{Function to validate output from lassosum.pipeline with external phenotype}
\usage{
validate(...)

\method{validate}{lassosum.pipeline}(ls.pipeline, test.bfile = NULL,
  keep = NULL, remove = NULL, pheno = NULL, covar = NULL,
  validate.function = function(x, y) cor(x, y, use = "complete"),
  trace = 1, destandardize = F, plot = T, exclude.ambiguous = T,
  cluster = NULL, rematch = !is.null(test.bfile), ...)
}
\arguments{
\item{...}{parameters to pass to \code{\link{sd.bfile}}}

\item{ls.pipeline}{A lassosum.pipeline object}

\item{test.bfile}{The (\href{https://www.cog-genomics.org/plink2/formats#bed}{PLINK bfile} for the test dataset}

\item{keep}{Participants to keep (see \code{\link{lassosum}} for more details)}

\item{remove}{Participants to remove}

\item{pheno}{A vector of phenotype OR a \code{data.frame} with 3 columns, the first 2 columns being headed "FID" and "IID", OR a filename for such a data.frame}

\item{covar}{A matrix of covariates OR a \code{data.frame} with 3 or more columns, the first 2 columns being headed "FID" and "IID", OR a filename for such a data.frame}

\item{validate.function}{Function with which to perform validation}

\item{trace}{Controls amount of output}

\item{destandardize}{Should coefficients from \code{\link{lassosum}} be 
destandardized using test dataset standard deviations before being returned?}

\item{plot}{Should the validation plot be plotted?}

\item{exclude.ambiguous}{Should ambiguous SNPs (C/G, A/T) be excluded?}

\item{cluster}{A \code{cluster} object from the \code{parallel} package for parallel computing}

\item{rematch}{Forces a rematching of the ls.pipline beta's with the new .bim file}
}
\description{
Function to validate output from lassosum.pipeline with external phenotype
}
\details{
Chooses the best \code{lambda} and \code{s} by validating 
polygenic score against an external phenotype in the testing dataset. 
If \code{pheno} is not specified, then the sixth column in the testing 
dataset \href{https://www.cog-genomics.org/plink2/formats#fam}{.fam}\code{.fam} file is used.
}
