% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.beta.R
\name{xp.beta}
\alias{xp.beta}
\title{Get the "overall" beta after cross-prediction}
\usage{
xp.beta(xp.plink.linear, xp.lassosum, type = c("Type1", "Type2"), load = "",
  save = "", force = FALSE)
}
\arguments{
\item{xp.plink.linear}{An object returned from xp.plink.linear()}

\item{xp.lassosum}{An object returned from xp.lassosum()}

\item{type}{Either the Type1 or Type2 Polygenic Score}

\item{load}{Load the pseudoinverse}

\item{save}{Save the pseudoinverse (in armadillo binary format)}
}
\value{
A vector of betas
}
