% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.cv.R
\name{xp.cv}
\alias{xp.cv}
\title{Perform cross-validation based on results from cross-prediction}
\usage{
xp.cv(xp.plink.linear, xp.lassosum, return.lpipe = FALSE, ...)
}
\arguments{
\item{xp.plink.linear}{An object returned from xp.plink.linear()}

\item{xp.lassosum}{An object returned from xp.lassosum()}

\item{return.lpipe}{For internal use}
}
\value{
A validate.lassosum.pipeline object
}
