% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.lassosum.list.R
\name{xp.lassosum.list}
\alias{xp.lassosum.list}
\title{For a list of xp.plink.linear objects}
\usage{
xp.lassosum.list(xp.plink.linear, trace = 0, keep.ref = NULL, ...)
}
\description{
For a list of xp.plink.linear objects
}
\keyword{internal}
