% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.lassosum.R
\name{xp.lassosum}
\alias{xp.lassosum}
\title{lassosum with cross-prediction}
\usage{
xp.lassosum(ss, LDblocks = ss$chr, ref.bfile = NULL, destandardize = TRUE,
  max.ref.bfile.n = 5000, details = FALSE, keep.ref = NULL,
  exclude.ambiguous = FALSE, validate.function = "cor", plot = TRUE,
  trace = 1, ...)
}
\arguments{
\item{ss}{An object from xp.plink.linear()}

\item{LDblocks}{LD blocks. See \code{\link{lassosum.pipeline}}}

\item{ref.bfile}{bfile of reference panel (if different from that used in \code{ss})}

\item{destandardize}{Should coefficients be destandardized}

\item{max.ref.bfile.n}{Maximum number of samples to use in ref.bfile (to improve speed)}

\item{details}{Should a detailed output be given?}

\item{keep.ref}{Participants to keep in the reference panel (see \code{parseselect} for expected input)}

\item{exclude.ambiguous}{Should ambiguous SNPs be excluded?}

\item{validate.function}{Function for validating polygenic score}

\item{plot}{Whether a validation plot should be drawn}

\item{...}{Other parameters to pass to lassosum.pipeline()}
}
\description{
We assume correlations are pre-calculated for the various 
folds, using PLINK or otherwise (cos PLINK is a lot faster).
}

