% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.lassosum.validate.R
\name{xp.lassosum.validate}
\alias{xp.lassosum.validate}
\title{For a list of xp.plink.linear objects}
\usage{
xp.lassosum.validate(list.of.lpipe, xp.plink.linear, Type2, pseudovalidation,
  scale = TRUE, plot, validate.function, cluster, trace, details,
  list.of.bfiles)
}
\description{
For a list of xp.plink.linear objects
}
\keyword{internal}
