% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.plink.linear.R
\name{xp.plink.linear}
\alias{xp.plink.linear}
\title{Generate summary statistics for cross-prediction}
\usage{
xp.plink.linear(bfile, nfolds = 5, fold = NULL, pheno = NULL,
  covar = NULL, keep = NULL, remove = NULL, extract = NULL,
  exclude = NULL, chr = NULL, force = FALSE, fast = TRUE, ...)
}
\arguments{
\item{nfolds}{Number of folds}

\item{fold}{A vector of fold ID}

\item{pheno}{A vector of phenotype}

\item{covar}{A vector/matrix of covariates}

\item{keep, remove, extract, exclude, chr}{see \code{parseselect}}
}
