% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.meta.R
\name{cp.meta}
\alias{cp.meta}
\title{Function to meta-analyse raw data summary statistics with external
summary statistics}
\usage{
cp.meta(xp.plink.linear, cor, chr = NULL, pos = NULL, snp = NULL,
  A1 = NULL, A2 = NULL, n = NULL, nonmiss = NULL,
  exclude.ambiguous = TRUE)
}
\arguments{
\item{xp.plink.linear}{An \code{xp.plink.linear} object}

\item{cor, chr, pos, snp, A1, A2}{see \code{lassosumm.pipeline}}

\item{n}{Sample size}

\item{nonmiss}{A vector of the number of non-missing observations}

\item{exclude.ambiguous}{Should ambiguous SNPs be excluded?}
}
\description{
Function to meta-analyse raw data summary statistics with external
summary statistics
}
\details{
This function performs a meta-analysis of the correlations 
coefficients as calculated in \code{xp.plink.linear} and some external 
correlations. If correlation coefficients are not available, these
can be converted from p-values using \code{p2cor}.
}
\note{
Either \code{n} or \code{nonmiss} must be specified
}
