% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.lassosum.pipeline.R
\name{merge.lassosum.pipeline}
\alias{merge.lassosum.pipeline}
\title{Merge a list of lassosum.pipeline objects}
\usage{
\method{merge}{lassosum.pipeline}(...)
}
\arguments{
\item{...}{\code{lassosum.pipeline} objects, separated by commas}
}
\value{
A \code{lassosum.pipeline} object
}
\description{
Merge a list of lassosum.pipeline objects
}
