% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseselect.R
\name{parseselect}
\alias{parseselect}
\title{Parse the keep/remove/extract/exclude/chr options}
\usage{
parseselect(
  bfile,
  extract = NULL,
  exclude = NULL,
  keep = NULL,
  remove = NULL,
  chr = NULL,
  export = FALSE,
  order.important = FALSE
)
}
\arguments{
\item{bfile}{plink file stem}

\item{extract}{SNPs to extract}

\item{exclude}{SNPs to exclude}

\item{keep}{samples to keep}

\item{remove}{samples to remove}

\item{chr}{a vector of chromosomes}

\item{export}{Logical. Whether to include the \code{bim} and/or \code{fam} data in the returned object.}

\item{order.important}{Logical. Whether to strictly enforce that the order of \code{extract} and/or \code{keep} must match with \code{bim} or \code{fam}, respectively.}
}
\value{
A list containing:
\item{keep}{Either NULL or a logical vector of which individuals to keep}
\item{extract}{Either NULL or a logical vector of which SNPs to extract}
\item{N}{Number of rows in the PLINK bfile}
\item{P}{Number of columns in the PLINK bfile}
\item{n}{Number of rows in the PLINK bfile after keep}
\item{p}{Number of columns in the PLINK bfile after extract}
\item{bfile}{File path to bfile stub}
\item{bimfile}{File path to bim file}
\item{famfile}{File path to fam file}
\item{bim}{Either \code{NULL} or a data frame of bim data}
\item{fam}{Either \code{NULL} or a data frame of fam data}
}
\description{
Parse the keep/remove/extract/exclude/chr options
}
\details{
\code{keep} and \code{remove} can take one of three
formats:
\enumerate{
\item A logical vector indicating which indivduals to keep/remove,
\item A \code{data.frame} with two columns giving the FID and IID of the indivdiuals to keep/remove (matching those in the .fam file), or
\item A character scalar giving the text file with the FID/IID.
}

Note that these files should not contain headers.

\code{extract} and \code{exclude} can be of format type \strong{(1)} or \strong{(3)} describe above, or a character vector of SNP ids (matching those in the .bim file).
}
