% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.linear.R
\name{plink.linear}
\alias{plink.linear}
\title{Obtain standardized coefficients from linear regression in PLINK}
\usage{
plink.linear(bfile, pheno, out = tempfile("lassosum.out"), keep = NULL,
  remove = NULL, extract = NULL, exclude = NULL, chr = NULL,
  covar = NULL, plink.cmd = "--linear standard-beta",
  ext = ".assoc.linear", ignore.stdout = TRUE, ...)
}
\arguments{
\item{keep, remove, extract, exclude, chr}{see parseselect()}
}
