% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp.plink.linear.R
\name{xp.plink.linear}
\alias{xp.plink.linear}
\title{Generate summary statistics for cross-prediction}
\usage{
xp.plink.linear(bfile, nfolds = 5, fold = NULL, pheno = NULL,
  covar = NULL, keep = NULL, remove = NULL, extract = NULL,
  exclude = NULL, chr = NULL, force = FALSE, fast = TRUE, ...)
}
\arguments{
\item{nfolds}{Number of folds}

\item{fold}{A vector of fold ID}

\item{pheno}{A vector of phenotype}

\item{covar}{A vector/matrix of covariates}

\item{keep, remove, extract, exclude, chr}{see \code{parseselect}}

\item{force}{Force}

\item{fast}{see details}
}
\details{
If fast == TRUE, the summary statistics are calculated for 
         each fold, and then combined together by averaging. For example,
         assume there are 5 folds, then for fold 1, the summary stats
         for fold 2, 3, 4, and 5 are averaged. If fast == FALSE, 
         we actually run plink using the entire dataset minus the fold 1, 
         and repeat this process for folds 2, 3, 4, and 5.
}
